requirements.txt
efficientnet_pytorch==0.7.1
einops==0.8.0
mamba_ssm==1.2.0.post1
matplotlib==3.8.0
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
scikit_learn==1.3.0
seaborn==0.13.2
timm==1.0.3
torch==2.2.0+cu121
torchsummary==1.5.1
torchview==0.2.6
torchvision==0.17.0
tqdm==4.65.0
split-folders==0.5.1